pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myagent"
version = "0.1.0"
description = "A minimal agent kernel scaffold inspired by nanobot-style modular design."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
myagent = "myagent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
